Changed 'or' label text value in query view
This is a more detailed description. WTF was the first thing that came
to my mind so then i decided to use this value for testing purposes.

Also the files in target\classes were generated by netbeans when i
opened the project. They seem empty, so i guess it's safe to commit
these as well.