Changed RsToRL to update schedule info in 1 go
Modified RestaurantDAO:
- Added setScheduleForMultipleRestaurants method. Given a list of
restaurants, for each restaurant we get the ID. For each id we get the
schedule from the DB. For each schedule taken from DB we update the
Restaurant object's schedule. We return back a list of restaurants with
updated schedules.
Modified RsToRL:
- When it creates a restaurant and adds it to the list, the schedule is
empty.
- After all the restaurants are created, we update their schedules using
a single connection, instead of making a new connection for each
restaurant to get the schedule
- Makes use of setScheduleForMultipleRestaurants method