Changed behavior closeView() Rest.EditController
When the closeView() method is called, the query view gets a refresh
instead of setting the visibility of the initial query view back to
true. This allows for the restaurants to be regenerated, instead of
loading the view back with the old restaurants data. This fixes the
problem with not having updated restaurant information.