Made a few changes to the program.
It was bugging me that on the filter side we have set areas and on the add restaurant part we had a text field. So i changed the areatextfield in the editrestaurant view. Everything seems to be working smoothly but please check the program and see.

Main Changes:
- RestaurantDetails.java
- RestaurantFactory.java
- RestaurantEditViewDataValidator.java
- ComponentValidator.java
- RestaurantEditController.java

Minor Changes:
- RestaurantSchedule.java

The major changes have to do with the dropdown included for the area.
The minor changes have to do with unnecessary text and imports.

Please let me know if you find any mistakes.